+++ src/main/java/ml/shifu/shifu/udf/FullScoreUDF.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Copyright [2012-2014] PayPal Software Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -61,7 +61,7 @@ public class FullScoreUDF extends AbstractTrainerUDF<Tuple> {
         tuple.append(cs.getMaxScore());
         tuple.append(cs.getMinScore());
 
-        for(Integer score: cs.getScores()) {
+        for(double score: cs.getScores()) {
             tuple.append(score);
         }
